Encode `Value` to `Json.Encode.Value`.
This allows us to parse broken JSON but then move back to the familiar world of
elm/json. It also means that, when an upstream source of broken JSON is fixed,
one can simply revert to using `Json.Decode.decodeString` with the same
decoders.

This also remove the query feature introduced in the previous commit. Using that
was cumbersome, and would bind people more closely to this module.